requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# AI/ML dependencies
langchain>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.0.40

# File handling
openpyxl>=3.1.0
python-dotenv>=1.0.0

# Web interface
streamlit>=1.28.0

# Data processing utilities
scipy>=1.11.0

# Additional dependencies for stability
requests>=2.28.0
typing-extensions>=4.0.0

PyPDF2
python-docx
fpdf
